LayoutTests/imported/w3c/web-platform-tests/streams/count-queuing-strategy.sharedworker-expected.txt

CONSOLE MESSAGE: line 10: ReferenceError: Can't find variable: SharedWorker

FAIL count-queuing-strategy.js shared worker wrapper file ReferenceError: Can't find variable: SharedWorker
